Fixed Resources to display correct status codes to run checks before running methods

+++ INFO310_project/src/main/java/resource/ProjectResource.java
@@ -10,6 +10,7 @@ import domain.ErrorMessage;
 import domain.Project;
 import io.jooby.Jooby;
 import io.jooby.StatusCode;
+import java.util.Collection;
 
 /**
  *
@@ -28,46 +29,64 @@ public class ProjectResource extends Jooby{
         path("/api/supervisor/projects/{staffID}", () -> {
             get("", ctx -> {
                 String id = ctx.path("staffID").value();
-                return dao.getProjectsByStaffID(id);
+                Collection<Project> projects = dao.getProjectsByStaffID(id);
+                if(projects.isEmpty()){
+                    return ctx.send(StatusCode.NOT_FOUND);
+                } else {
+                    return projects;
+                }
             });
         });
         
         path("/api/projects/{projectID}", () -> {
             get("", ctx -> {
                 String id = ctx.path("projectID").value();
-                return dao.getProjectByID(id);
+                Project project = dao.getProjectByID(id);
+                if(project == null){
+                    return ctx.send(StatusCode.NOT_FOUND);
+                } else {
+                    return project;
+                }
             });
             
             put("", ctx -> {
                 String id = ctx.path("projectID").value();
-                Project project = ctx.body().to(Project.class);
-                if(!id.equals(project.getProjectID())){
-                    return ctx
-                            .setResponseCode(StatusCode.CONFLICT)
-                            .render(new ErrorMessage("Modifying the project's ID via this operation is not allowed.  Create a new project instead."));
+                if(dao.getProjectByID(id) == null){
+                    return ctx.send(StatusCode.NOT_FOUND);
                 } else {
-                    dao.updateProjectByID(id, project);
-                    return ctx.send(StatusCode.NO_CONTENT);
+                    Project project = ctx.body().to(Project.class);
+                    if(!id.equals(project.getProjectID())){
+                        return ctx
+                                .setResponseCode(StatusCode.CONFLICT)
+                                .render(new ErrorMessage("Modifying the project's ID via this operation is not allowed.  Create a new project instead."));
+                    } else {
+                        dao.updateProjectByID(id, project);
+                        return ctx.send(StatusCode.NO_CONTENT);
+                    }
                 }
             });
             
             delete("", ctx -> {
                 String id = ctx.path("projectID").value();
-                dao.deleteProjectByID(id);
-                return ctx.send(StatusCode.NO_CONTENT);
+                if(dao.getProjectByID(id) == null){
+                    return ctx.send(StatusCode.NOT_FOUND);
+                } else {
+                    dao.deleteProjectByID(id);
+                    return ctx.send(StatusCode.NO_CONTENT);
+                }
             });
         });
         
         path("/api/supervisor/projects", () -> {
             post("", ctx -> {
                 Project project = ctx.body().to(Project.class);
-                if (ProjectCollectionsDao.exists(project.getProjectID())) {
+                if (dao.getProjectByID(project.getProjectID()) == null) {
+                    dao.saveProject(project);
+                    return ctx.send(StatusCode.CREATED);
+                } else {
                     return ctx
                             .setResponseCode(StatusCode.CONFLICT)
                             .render(new ErrorMessage("There is already an existing project with this ID in the system"));
-                } else {
-                    dao.saveProject(project);
-                    return ctx.send(StatusCode.CREATED);
                 }
             });
         });
