+++ INFO310_project/src/main/java/resource/SupervisorResource.java
@@ -4,7 +4,6 @@
  */
 package resource;
 
-import dao.SupervisorCollectionsDao;
 import dao.SupervisorDao;
 import domain.ErrorMessage;
 import domain.Supervisor;
@@ -27,25 +26,39 @@ public class SupervisorResource extends Jooby{
 
             get("", ctx -> {
                 String id = ctx.path("staffID").value();
-                return dao.getSupervisorById(id);
+                Supervisor supervisor = dao.getSupervisorById(id);
+                if(supervisor == null){
+                    return ctx.send(StatusCode.NOT_FOUND);
+                } else {
+                    return supervisor;
+                }
             });
 
             delete("", ctx -> {
                 String id = ctx.path("staffID").value();
-                dao.deleteSupervisor(id);
-                return ctx.send(StatusCode.NO_CONTENT);
+                if(dao.getSupervisorById(id) == null){
+                    return ctx.send(StatusCode.NOT_FOUND);
+                } else {
+                    dao.deleteSupervisor(id);
+                    return ctx.send(StatusCode.NO_CONTENT);
+                }
             });
 
             put("", ctx -> {
                 String id = ctx.path("staffID").value();
-                Supervisor supervisor = ctx.body().to(Supervisor.class);
-                if (!id.equals(supervisor.getStaffID())) {
-                    return ctx
-                            .setResponseCode(StatusCode.CONFLICT)
-                            .render(new ErrorMessage("Modifying the product's ID via this operation is not allowed.  Create a new product instead."));
-                } else {
-                    dao.updateSupervisor(id, supervisor);
-                    return ctx.send(StatusCode.NO_CONTENT);
+                if(dao.getSupervisorById(id) == null){
+                    return ctx.send(StatusCode.NOT_FOUND);
+                }
+                else{
+                    Supervisor supervisor = ctx.body().to(Supervisor.class);
+                    if (!id.equals(supervisor.getStaffID())) {
+                        return ctx
+                                .setResponseCode(StatusCode.CONFLICT)
+                                .render(new ErrorMessage("Modifying the product's ID via this operation is not allowed.  Create a new product instead."));
+                    } else {
+                        dao.updateSupervisor(id, supervisor);
+                        return ctx.send(StatusCode.NO_CONTENT);
+                    }
                 }
             });
         });  
@@ -53,13 +66,13 @@ public class SupervisorResource extends Jooby{
         path("/api/sign-up/supervisor", () -> {
             post("", ctx -> {
                 Supervisor supervisor = ctx.body().to(Supervisor.class);
-                if (SupervisorCollectionsDao.exists(supervisor.getStaffID())) {
-                    return ctx
-                            .setResponseCode(StatusCode.CONFLICT)
-                            .render(new ErrorMessage("There is already an existing Staff member with this ID in the system"));
-                } else {
+                if (dao.getSupervisorById(supervisor.getStaffID()) == null) {
                     dao.saveSupervisor(supervisor);
                     return ctx.send(StatusCode.CREATED);
+                } else {
+                  return ctx
+                            .setResponseCode(StatusCode.CONFLICT)
+                            .render(new ErrorMessage("There is already an existing Staff member with this ID in the system"));  
                 }
             });
         });
